requirements.txt
numpy
PyOpenGL
cyglfw3
six